move client to separate directory
