+++ benchmarks/generator/utils.py
@@ -254,23 +254,3 @@ def load_config(config_path: str) -> Dict[str, Any]:
         config = json.load(file)
     return config
 
-def load_workload(input_path: str) -> List[Any]:
-    load_struct = None
-    if input_path.endswith(".jsonl"):
-        with open(input_path, "r") as file:
-            load_struct = [json.loads(line) for line in file]
-    else:
-        with open(input_path, "r") as file:
-            load_struct = json.load(file)
-    return load_struct
-
-# Function to wrap the prompt into OpenAI's chat completion message format.
-def wrap_prompt_as_chat_message(prompt: str):
-    """
-    Wrap the prompt into OpenAI's chat completion message format.
-
-    :param prompt: The user prompt to be converted.
-    :return: A list containing chat completion messages.
-    """
-    user_message = {"role": "user", "content": prompt}
-    return [user_message]
